Move asynchronous handler logic to a separate file
I want to build on top of this logic to add more complex logic that
would also include retries. Separating all of the async management logic
into a single file helps to encapsulate this complexity.
